chore: rename project to platform-java
Rename the package, project, repo to platform-java.
