+++ CONTRIBUTING.md
@@ -80,6 +80,6 @@ trace the normal flow of execution in production code.
 
 ## License
 
-By contributing to Toolbox Java, you agree that your contributions will be licensed under the
+By contributing to Platform Java, you agree that your contributions will be licensed under the
 [Apache 2.0 License](https://www.apache.org/licenses/LICENSE-2.0). A copy of the license is
 available in the [LICENSE.md](LICENSE.md) file in the root directory of this source tree.
