+++ src/main/java/com/reactivemarkets/platform/bsp/BspHandler.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.reactivemarkets.toolbox.bsp;
+package com.reactivemarkets.platform.bsp;
 
 import io.netty.buffer.ByteBuf;
