+++ src/main/java/com/reactivemarkets/platform/util/LoggerUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.reactivemarkets.toolbox.util;
+package com.reactivemarkets.platform.util;
 
 import org.apache.log4j.ConsoleAppender;
 import org.apache.log4j.EnhancedPatternLayout;
